move tests.pp to tests.__init__
