+++ lib/iris/tests/test_cdm.py
@@ -23,7 +23,6 @@ import iris.coords
 import iris.cube
 import iris.fileformats
 import iris.fileformats.dot
-import iris.tests.pp as pp
 import iris.tests.stock
 
 
@@ -1287,7 +1286,7 @@ class TestTrimAttributes(tests.IrisTest):
 
 
 @tests.skip_data
-class TestMaskedData(tests.IrisTest, pp.PPTest):
+class TestMaskedData(tests.IrisTest, tests.PPTest):
     def _load_3d_cube(self):
         # This 3D data set has a missing a slice with SOME missing values.
         # The missing data is in the pressure = 1000 hPa, forcast_period = 0,
